+++ components/CustomMeta/index.ts
@@ -1,0 +1,1 @@
+export { default as CustomMeta } from "./CustomMeta";
