+++ components/DefaultMeta/index.ts
@@ -1,2 +1,1 @@
-export { default as CustomMeta } from "./CustomMeta";
 export { default as DefaultMeta } from "./DefaultMeta";
