+++ components/Notes/index.ts
@@ -1,0 +1,1 @@
+export { default as Notes } from "./Notes";
