--- components/RecentPosts/index.ts
@@ -1,1 +1,0 @@
-export { default as RecentPosts } from "./RecentPosts";
