+++ components/index.ts
@@ -1,7 +1,8 @@
+import { DefaultMeta } from "./DefaultMeta";
+
 // Card
 export { Card } from "./Card";
 export { ProjectCard } from "./ProjectCard";
-export { RecentPosts } from "./RecentPosts";
 
 // Link
 export { CustomLink } from "./CustomLink";
@@ -14,20 +15,20 @@ export { Footer } from "./Footer";
 
 // Section
 export { MdxSection } from "./MdxSection";
-export { NoteSection } from "./NoteSection";
 
 // Button
 export { DarkModeButton } from "./DarkModeButton";
 export { ToTopButton } from "./ToTopButton";
 
 // TOC
-export { TOC } from "./TOC";
+export { Toc } from "./Toc";
 
 // Meta
-export { CustomMeta, DefaultMeta } from "./Meta";
+
+export { DefaultMeta } from "./DefaultMeta";
+export { CustomMeta } from "./CustomMeta";
 
 // Filter
-export { Tags } from "./Tags";
 export { NoteFilter } from "./NoteFilter";
 
 // Icons
@@ -38,3 +39,5 @@ export { AppLayout, PageLayout, MdxLayout } from "./Layout";
 
 // Input
 export { InputWithRef } from "./Input";
+
+export { Notes } from "./Notes";
