--- components/noteSection/index.ts
@@ -1,1 +1,0 @@
-export { default as NoteSection } from "./NoteSection";
