--- components/noteSection/types.ts
@@ -1,7 +1,0 @@
-export type Props = {
-  created_time: string;
-  description: string;
-  id?: string;
-  title: string;
-  url: string;
-};
