--- components/toc/types.tsx
@@ -1,5 +1,0 @@
-export type HeadingType = {
-  slug: string;
-  heading: "heading2" | "heading3" | "heading4";
-  text: string;
-};
