+++ constants/constants.ts
@@ -2,7 +2,7 @@ export const NAV_PATHS = [
   { path: "/post", title: "POST" },
   { path: "/projects", title: "PROJECTS" },
   { path: "/search", title: "SEARCH" },
-  { path: "/note", title: "NOTES" },
+  { path: "/note", title: "NOTES/React" },
 ];
 
 export const NOTE_FILTER_OPTIONS = [
