+++ hooks/index.ts
@@ -4,6 +4,6 @@ export { default as useInput } from "./useInput";
 export { default as useInfiniteScroll } from "./useInfiniteScroll";
 export { default as useSavedScroll } from "./useSavedScroll";
 export { default as useSessionStorage } from "./useSessionStorage";
-export { default as useTocHighLight } from "./useTocHighLight";
 export { default as useClose } from "./useClose";
 export { default as useDarkMode } from "./useDarkMode";
+export { default as useToc } from "./useToc";
