+++ pages/_app.tsx
@@ -2,8 +2,8 @@ import "../styles/globals.css";
 import { Fragment } from "react";
 
 import type { AppProps } from "next/app";
-import DefaultMeta from "components/Meta/DefaultMeta";
-import { AppLayout } from "components";
+
+import { AppLayout, DefaultMeta } from "components";
 import useSavedScroll from "hooks/useSavedScroll";
 
 function MyApp({ Component, pageProps }: AppProps) {
