+++ pages/index.tsx
@@ -5,7 +5,7 @@ import type { NextPage } from "next";
 import { allPosts, Post as PostType } from "contentlayer/generated";
 
 // components
-import { Card, RecentPosts } from "components";
+import { Card } from "components";
 
 type Props = {
   featuredPost: PostType;
@@ -23,7 +23,32 @@ const Home: NextPage<Props> = ({ featuredPost, recentPosts }) => {
         slug={featuredPost.path}
         cardType='featuredCard'
       />
-      <RecentPosts recentPosts={recentPosts} />
+      <div className='mx-auto mt-10 md:max-w-[70%]'>
+        <h2 className='mb-8 text-center text-3xl font-bold italic md:mb-10'>
+          RECENT POSTS
+        </h2>
+        {recentPosts.map(
+          ({
+            title,
+            description,
+            publishedAtFormatted,
+            thumbnailImg,
+            path,
+            tags,
+          }: PostType) => (
+            <Card
+              title={title}
+              description={description}
+              publishedAt={publishedAtFormatted}
+              thumbnailImg={thumbnailImg}
+              key={title}
+              slug={path}
+              tags={tags}
+              cardType='verticalCard'
+            />
+          ),
+        )}
+      </div>
     </Fragment>
   );
 };
