+++ pages/post/[slug].tsx
@@ -3,7 +3,7 @@ import { Fragment } from "react";
 import type { GetStaticProps, InferGetStaticPropsType } from "next";
 import type { Params } from "types/types";
 
-import { MdxSection, TOC, CustomMeta, MdxLayout } from "components";
+import { MdxSection, MdxLayout, CustomMeta, Toc } from "components";
 import { allPosts } from "contentlayer/generated";
 
 const Blog = ({ data }: InferGetStaticPropsType<typeof getStaticProps>) => {
@@ -20,7 +20,7 @@ const Blog = ({ data }: InferGetStaticPropsType<typeof getStaticProps>) => {
       />
       <MdxLayout>
         <MdxSection postData={data} />
-        <TOC headings={headings} title={title} />
+        <Toc title={title} headings={headings} />
       </MdxLayout>
     </Fragment>
   );
