+++ pages/projects/index.tsx
@@ -1,36 +1,27 @@
-// components
-import { PageLayout, ProjectCard } from "components";
+import type { GetStaticProps, InferGetStaticPropsType } from "next";
 
-// mdx
+import { ProjectCard } from "components";
+
 import { allProjects, Project as ProjectType } from "contentlayer/generated";
 
-type Props = {
-  allProjects: ProjectType[];
-};
-
-const Projects: React.FC<Props> = ({ allProjects }) => {
+const Projects = ({ allProjects }: InferGetStaticPropsType<GetStaticProps>) => {
   return (
-    <PageLayout
-      title='PROJECTS...'
-      description='프로젝트를 진행하면서 이런 생각을 했습니다. 🤡'
-    >
-      <div className='flex flex-col md:flex-row md:items-center md:justify-center md:space-x-5'>
-        {allProjects.map(({ title, thumbnailImg, slug }: ProjectType) => (
-          <ProjectCard
-            key={title}
-            title={title}
-            thumbnailImg={thumbnailImg}
-            slug={slug}
-          />
-        ))}
-      </div>
-    </PageLayout>
+    <div className='flex min-h-[720px] flex-col px-5 md:flex-row md:items-center md:justify-center md:space-x-3'>
+      {allProjects.map(({ title, thumbnailImg, slug }: ProjectType) => (
+        <ProjectCard
+          key={title}
+          title={title}
+          thumbnailImg={thumbnailImg}
+          slug={slug}
+        />
+      ))}
+    </div>
   );
 };
 
 export default Projects;
 
-export const getStaticProps = async () => {
+export const getStaticProps: GetStaticProps = async () => {
   return {
     props: {
       allProjects,
